Fixing DB errors and updating db schema
git-tfs-id: [https://tfs.codeplex.com/tfs/tfs04]$/polaris;C38292

+++ Main/Polaris.Dbl/Schema Objects/Tables/Polaris.PlayLog.table.sql
@@ -1,8 +1,8 @@
 ﻿CREATE TABLE [Polaris].[PlayLog]
 (
-[PlayLogId] [int] NOT NULL IDENTITY(1, 1),
-[GameId] [int] NOT NULL,
-[UserId] [int] NOT NULL,
+[PlayLogId] [bigint] NOT NULL IDENTITY(1, 1),
+[GameId] [bigint] NOT NULL,
+[UserId] [bigint] NOT NULL,
 [Date] [datetime] NOT NULL,
 [Score] [decimal] (18, 0) NOT NULL
 ) ON [PRIMARY]
